requirements.txt
fastapi
uvicorn
elasticsearch
google-cloud-vision
google-cloud-language
google-cloud-firestore
google-cloud-storage
google-cloud-bigquery
firebase-admin
python-dotenv
